+++ pgcom/__init__.py
@@ -1,3 +1,4 @@
+from .connector import *  # noqa: F401, F403
 from .commuter import *  # noqa: F401, F403
 from .listener import *  # noqa: F401, F403